requirements.txt
asgiref==3.3.4
Django==3.2
django-crispy-forms==1.11.2
Pillow==8.2.0
python-dotenv==0.17.0
pytz==2021.1
sqlparse==0.4.1
